Move schema validation, omit error data on logs
