+++ src/datasources/config-api/config-api.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ConfigApi } from './config-api.service';
 import { CacheFirstDataSourceModule } from '../cache/cache.first.data.source.module';
 import { ValidationErrorFactory } from '../errors/validation-error-factory';
-import { JsonSchemaService } from '../common/json-schema.service';
+import { JsonSchemaService } from '../../common/schema/json-schema.service';
 
 @Module({
   imports: [CacheFirstDataSourceModule],
