+++ src/datasources/exchange-api/exchange-api.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { HttpErrorFactory } from '../errors/http-error-factory';
 import { ExchangeApi } from './exchange-api.service';
 import { ValidationErrorFactory } from '../errors/validation-error-factory';
-import { JsonSchemaService } from '../common/json-schema.service';
+import { JsonSchemaService } from '../../common/schema/json-schema.service';
 
 @Module({
   providers: [
