+++ src/datasources/transaction-api/transaction-api.module.ts
@@ -3,7 +3,7 @@ import { ConfigApiModule } from '../config-api/config-api.module';
 import { TransactionApiManager } from './transaction-api.manager';
 import { CacheFirstDataSourceModule } from '../cache/cache.first.data.source.module';
 import { ValidationErrorFactory } from '../errors/validation-error-factory';
-import { JsonSchemaService } from '../common/json-schema.service';
+import { JsonSchemaService } from '../../common/schema/json-schema.service';
 
 @Module({
   imports: [ConfigApiModule, CacheFirstDataSourceModule],
