+++ src/datasources/transaction-api/transaction-api.service.spec.ts
@@ -6,7 +6,7 @@ import { ValidationErrorFactory } from '../errors/validation-error-factory';
 import { HttpException } from '@nestjs/common';
 import backboneFactory from '../../chains/entities/__tests__/backbone.factory';
 import { balanceFactory } from './entities/__tests__/balance.factory';
-import { JsonSchemaService } from '../common/json-schema.service';
+import { JsonSchemaService } from '../../common/schema/json-schema.service';
 
 const BALANCES: Balance[] = [balanceFactory(), balanceFactory()];
 const BACKBONE: Backbone = backboneFactory();
